+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 - Show usernames in normal letter case
 - Show event author when viewing specific event
 - Show message if a post or comment has been edited
+- Add button to create a new post to each forum
+- Remove button to create a new post from the sidebar